challenge-admin/requirements.txt
Babel==1.3
Flask==0.10.1
Flask-Babel==0.9
Flask-Cache==0.13.1
Flask-Mandrill==0.2
Flask-RESTful==0.3.0
Flask-SQLAlchemy==2.0
Flask-WTF==0.10.3
Jinja2==2.7.3
MarkupSafe==0.23
PypeR==1.1.2
SQLAlchemy==0.9.8
WTForms==2.0.1
Werkzeug==0.9.6
aniso8601==0.85
argparse==1.2.1
itsdangerous==0.24
mongokit==0.9.1.1
mysql-connector-python==2.0.2
py-bcrypt==0.4
pymongo==2.7.2
python-memcached==1.53
pytz==2014.9
requests==2.5.1
sendgrid==1.4.0
six==1.8.0
smtpapi==0.2.0
speaklater==1.3
wsgiref==0.1.2